requirements.txt
aiogram==3.*
python-dotenv==0.21.0
yandex-cloud-ml-sdk==0.3.0
aiohttp~=3.11.11
aiofiles~=24.1.0
matplotlib~=3.9.2